+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 ### Breaking
-Recast "input" as "arg parser"
+Refactor:
+- "input" --> "arg parser" 
+- "getValue" --> "parse"
 ## 0.4.0
 ### Breaking
 Numerous tweaks to make the API more expressive and grepable, e.g. CliLeaf instead of just Leaf. The docs are still a work-in-progress but the TypeScript types are spot-on. Let them be your guide!